[Refactoring] Move ExpandTernaryExpr to its own file
